+++ pda/pdtree.cpp
@@ -163,7 +163,7 @@ void PDTree::readParams(Params &params) {
 	LeafMapName lsn;
 	buildLeafMapName(lsn);
 	tax_weight.resize((unsigned long) ntaxa, 0);
-	for (int i = 0; i < tax_name.size(); i++) {
+	for (size_t i = 0; i < tax_name.size(); i++) {
 		LeafMapName::iterator nameit = lsn.find(tax_name[i]);
 		if (nameit == lsn.end())
 			outError(ERR_NO_TAXON, tax_name[i]);
